requirements.txt
antiorm==1.2.1
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
db==0.1.1
db-sqlite3==0.0.1
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
jsmin==2.2.2
psycopg2==2.9.1
psycopg2-binary==2.8.6
Pygments==2.9.0
PySocks==1.7.1
pytz==2020.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.3.1
urllib3==1.26.6
whitenoise==5.2.0
